feat(eigenstructure): Phase 4 — full eigenvalue study across 8 datasets
Runs the Phase 3 v2 surrogate infrastructure (Fréchet-mean common
frame, permutation-null calibration, four estimator variants) across
the real-data benchmarks that will carry the paper's empirical
section. Scope and rationale in
``docs/plans/2026-04-19-phase4-final-eigenvalue-study.md``.

Datasets (8 total): four synthetic SBM at ``diversity ∈ {0, 0.33, 0.67,
1.0}``, ``spectre_sbm`` (upstream DiGress fixture, 68/200 graphs in
the 30-80 filter band), ``enzymes`` (305/600 in 30-60), ``proteins``
(383/1113 in 30-80), and ``collab`` (3715/5000 in 30-80). COLLAB newly
wired into ``PyGDatasetWrapper`` via a 10-line patch.

Size filtering (not zero-padding) keeps each dataset at a fixed ``n``
so the Fréchet mean has a well-defined ambient dimension. Retention
rates (34 % for proteins, 51 % for enzymes, 74 % for collab) are
documented in the report; the limitation is flagged explicitly.

Sweep: 8 datasets × 5 seeds × 2 frame modes × 2 noise types × 5 noise
levels × 4 estimators × 4 k × 2 (real + null) = **25 600 cells**,
completed in ~11 min local CPU. Double-precision eigh throughout
(float32 failed to converge on ENZYMES' near-degenerate bio graphs at
``ε = 0.01``; fixed both in the noised collector and the Phase 4 runner).

Headline at (frechet, knn_top_k, k=8, gaussian, ε=0.1):

  1. sbm_d1.00 margin 0.629 ✓
  2. collab    margin 0.545 ✓
  3. sbm_d0.67 margin 0.459 ✓
  4. spectre   margin 0.272 ✓
  5. sbm_d0.33 margin 0.205 ✓
  6. proteins  margin 0.184 ✓
  7. enzymes   margin 0.100 ✓
  8. sbm_d0.00 margin 0.043 ✗

**7 / 8 pass the ≥ 0.10 calibrated-margin threshold.** The one failure
is sbm_d0.00, the zero-diversity null case — exactly where the Phase 3
v2 diversity knob was designed to produce a null result. Passing here
would have invalidated the whole methodology.

Stability across the full (noise_type × ε × estimator × k) grid = 160
cells: the **top-2 set {sbm_d1.00, collab} matches the headline in
160/160 cells (100 %)** — the strongest real-data win. Top-1 and
mid-tier rankings shuffle within tolerance (12 % and 4 % respectively)
because marginal datasets' calibrated margins cluster in the 0.10–0.30
band.

Noise-level sensitivity: margins decline modestly as ``ε`` grows (top
eigenvalues get noisier → harder to condition on). Enzymes dips below
the 0.10 pass bar at ``ε = 0.2`` (0.08); every other dataset stays
well above threshold across the full noise range.

Synthetic-SBM diversity monotonicity reproduces inside the Phase 4
grid: margin 0.04 → 0.21 → 0.46 → 0.63 across diversity
∈ {0, 0.33, 0.67, 1.0}, matching Phase 3 v2 cleanly at ε=0.1, k=8.

Ready for paper-figure extraction from the 25 600-row CSV at
``docs/reports/2026-04-19-phase4-eigenvalue-study/phase4_sweep.csv``.

+++ tmgg/src/tmgg/experiments/eigenstructure_study/noised_collector.py
@@ -441,13 +441,22 @@ class NoisedEigenstructureCollector:
                 A = A_noised[i]
                 L = compute_laplacian(A)
 
-                vals_a, vecs_a = torch.linalg.eigh(A.float())
-                vals_l, vecs_l = torch.linalg.eigh(L.float())
+                # Noised adjacency matrices can be numerically
+                # ill-conditioned or near-degenerate — especially on real
+                # bio benchmarks like ENZYMES where the underlying graph
+                # has high symmetry and small perturbations keep
+                # eigenvalues effectively repeated. float32 eigh fails to
+                # converge on those graphs (observed in the Phase 4 sweep
+                # on ENZYMES at ε=0.01). Run eigh in float64 and downcast
+                # for storage — precision is cheap, a crash here kills
+                # the whole sweep.
+                vals_a, vecs_a = torch.linalg.eigh(A.double())
+                vals_l, vecs_l = torch.linalg.eigh(L.double())
 
-                eig_adj_vals_list.append(vals_a)
-                eig_adj_vecs_list.append(vecs_a)
-                eig_lap_vals_list.append(vals_l)
-                eig_lap_vecs_list.append(vecs_l)
+                eig_adj_vals_list.append(vals_a.float())
+                eig_adj_vecs_list.append(vecs_a.float())
+                eig_lap_vals_list.append(vals_l.float())
+                eig_lap_vecs_list.append(vecs_l.float())
 
             eig_adj_vecs = torch.stack(eig_adj_vecs_list)
